Weow.

+++ src/main/kotlin/com/forcetower/sagres/cookies/SerializableCookie.kt
@@ -27,7 +27,6 @@ import java.io.ObjectInputStream
 import java.io.ObjectOutputStream
 import java.io.Serializable
 import okhttp3.Cookie
-import okhttp3.internal.and
 
 class SerializableCookie : Serializable {
     @Transient
@@ -133,7 +132,7 @@ class SerializableCookie : Serializable {
         private fun byteArrayToHexString(bytes: ByteArray): String {
             val sb = StringBuilder(bytes.size * 2)
             for (element in bytes) {
-                val v = element and 0xff
+                val v = element.toInt() and 0xff
                 if (v < 16) {
                     sb.append('0')
                 }
